+++ webextension/action.js
@@ -33,9 +33,7 @@ document.getElementById('pin').onclick = function() {
 };
 
 document.getElementById('capture').onclick = function() {
-	getTab().then(function(tab) {
-		chrome.tabs.executeScript(tab.id, {file: 'thumbnail.js'}, () => {
-			window.close();
-		});
+	chrome.runtime.sendMessage({name: 'Thumbnails.capture'}, () => {
+		window.close();
 	});
 };